requirements.txt
discord.py~=1.5.1
emoji~=0.6.0